requirements-base.txt
requests == 2.7.0
colorama >= 0.3.3
pyquery >= 1.2.9
pytz==2016.10
lxml >= 3.4.4
pyyaml >= 3.13
flask == 0.10.1
prompt_toolkit == 2.0.3
terminaltables == 3.1.0
Pygments == 2.1.3
wcwidth == 0.1.7
six >= 1.10.0
boto3 >= 1.4.7
python-gnupg
watchdog==0.8.3
keyring==16.0.1
